FastAPI/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
anyio=3.7.0=pypi_0
apscheduler=3.10.1=pypi_0
async-generator=1.10=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py310haa95532_0
cffi=1.15.1=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
dnspython=2.3.0=pypi_0
email-validator=2.0.0.post2=pypi_0
exceptiongroup=1.1.1=pypi_0
fastapi=0.95.2=pypi_0
flask=2.2.3=pypi_0
frozenlist=1.3.3=pypi_0
future=0.18.3=pypi_0
h11=0.14.0=pypi_0
httpcore=0.17.2=pypi_0
httptools=0.5.0=pypi_0
httpx=0.24.1=pypi_0
idna=3.4=pypi_0
iottalk-paho-mqtt=1.4.0.dev3=pypi_0
iottalk-py=2.3.1=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=pypi_0
libffi=3.4.2=hd77b12b_6
line-bot-sdk=2.4.2=pypi_0
lxml=4.9.2=pypi_0
markupsafe=2.1.2=pypi_0
multidict=6.0.4=pypi_0
numpy=1.24.3=pypi_0
openssl=1.1.1t=h2bbff1b_0
orjson=3.9.0=pypi_0
outcome=1.2.0=pypi_0
pandas=2.0.0=pypi_0
pip=23.0.1=py310haa95532_0
pycparser=2.21=pypi_0
pydantic=1.10.8=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.9=h966fe2a_2
python-dateutil=2.8.2=pypi_0
python-dotenv=1.0.0=pypi_0
python-multipart=0.0.6=pypi_0
pytz=2023.3=pypi_0
pyyaml=6.0=pypi_0
requests=2.28.2=pypi_0
selenium=4.9.0=pypi_0
setuptools=65.6.3=py310haa95532_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.1=h2bbff1b_0
starlette=0.27.0=pypi_0
tk=8.6.12=h2bbff1b_0
tqdm=4.65.0=pypi_0
trio=0.22.0=pypi_0
trio-websocket=0.10.2=pypi_0
typing-extensions=4.6.3=pypi_0
tzdata=2023.3=pypi_0
tzlocal=5.0.1=pypi_0
ujson=5.7.0=pypi_0
urllib3=1.26.15=pypi_0
uvicorn=0.22.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchfiles=0.19.0=pypi_0
websockets=11.0.3=pypi_0
werkzeug=2.2.3=pypi_0
wheel=0.38.4=py310haa95532_0
wincertstore=0.2=py310haa95532_2
wsproto=1.2.0=pypi_0
xz=5.2.10=h8cc25b3_1
yarl=1.9.2=pypi_0
zlib=1.2.13=h8cc25b3_0
